requirements.txt
alembic==1.13.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1

Jinja2==3.1.3

psycopg2-binary==2.9.9
psycopg2-pool==1.2

python-dotenv==1.0.1
python-jose==3.3.0

SQLAlchemy==2.0.27

Werkzeug==3.0.1

gunicorn==22.0.0